Improve Encoder inheritance
- Stick common code into the base class as an abstract method
- Base class implements new IEncoder interface
- Add _allowNumbers variable
