+++ src/encoder/Encoder.cs
@@ -2,9 +2,10 @@
 
 namespace PitchConverter.Encoder
 {
-    public abstract class Encoder
+    public abstract class Encoder : IEncoder
     {
         private protected string _pitchClasses;
+        private protected bool _allowNumbers;
         public int StartOctave { get; set; }
         public bool IncludeRests { get; set; }
         public OutputFormat Format { get; init; }
@@ -12,26 +13,31 @@ namespace PitchConverter.Encoder
         public Encoder(OutputFormat format)
         {
             _pitchClasses = PitchConstants.PitchClassSets.NonChromaticPitchClasses;
+            _allowNumbers = true;
             StartOctave = 4;
             IncludeRests = true;
             Format = format;
         }
 
-        public abstract List<MusicSymbol> Encode(string input);
+        public virtual List<MusicSymbol> Encode(string input)
+        {
+            List<MusicSymbol> music = new();
+            for (int i = 0; i < input.Length; i++)
+            {
+                MusicSymbol ms = CharConverter.CharToPitch(input[i], _pitchClasses, StartOctave, _allowNumbers);
+                if (IncludeRests || ms.GetPitchClass() != -1)
+                {
+                    music.Add(ms);
+                }
+            }
+            return music;
+        }
 
-        public virtual void EncodeToFile(string input, OutputFormat outputFormat)
+        public void EncodeToFile(string input, OutputFormat outputFormat)
         {
             List<MusicSymbol> music = Encode(input);
             // Do file writing here
             throw new NotImplementedException();
         }
-
-        public enum OutputFormat
-        {
-            Text,
-            Midi,
-            MusicXml
-        }
-
     }
 }
