+++ src/encoder/IEncoder.cs
@@ -1,0 +1,10 @@
+﻿using PitchConverter.Pitch;
+
+namespace PitchConverter.Encoder
+{
+    public interface IEncoder
+    {
+        List<MusicSymbol> Encode(string input);
+        void EncodeToFile(string input, OutputFormat outputFormat);
+    }
+}
