+++ src/encoder/OutputFormat.cs
@@ -1,0 +1,9 @@
+﻿namespace PitchConverter.Encoder
+{
+    public enum OutputFormat
+    {
+        Text,
+        MusicXml,
+        Midi
+    }
+}
